10/10 기능수정 1. 모든 Singleton클래스들을 일반 MonoBehaviour로 바꾸고 참조가 필요한 경우 변수로 들고 있게 하였다 => 어차피 참조하는곳들은 여러번 쓰이므로, 싱글톤일때의 문제점들이 사라졌다. 2. Enemy를 Pool로 쓰는데 일반 Pool과 EnemyPool을 분리해주었고 PoolObjectManager에서 함수를 오버로드해주었다. => 현재 PoolObject의 키가 Enum이기 때문에 Enemy만 따로 분리해 사용하는게 편하기 때문에 3. 모든 타워는 스폰할 때 DoTween을 이용해 약간 위에서 떨어지듯 스폰되게 연출해주었다. 4. TargetingTower들을 각자 공격에 맞는 Tween으로 움직임을 주었다. 5. UnitTower에서 유닛이 리스폰 될 때 남은 시간을 ProgressBar로 보여준다.

+++ TowerDefense/Assets/Scripts/ProjectileControl/CanonProjectile.cs
@@ -62,7 +62,7 @@ namespace ProjectileControl
             }
         }
 
-        public override void Init(int dmg, Transform t)
+        public override void Init(ushort dmg, Transform t)
         {
             base.Init(dmg, t);
             Physics.Raycast(t.position + t.forward * 2 + Vector3.up * 2, Vector3.down, out var hit, 10);
